chore(size-guard): split the two files the formatter itself pushes over 51200
Both are pure code motion with no behaviour change, and both follow the house
pattern (`admission_journal_tests.rs` in #2589, `build_lease_cap_arming_tests.rs`
in #2593): split, never `djinn:allow-oversize`. An oversize marker on a file the
project's own formatter tips past the ceiling does not remove the trap, it just
moves it to whoever edits the file next.

`djinn-db`'s `test_support.rs` was 1428 lines but 51209 BYTES — nine over, with
the line count comfortably inside its limit, which is why the failure reads as
inexplicable. `cargo fmt` alone is enough to produce it, so `main` has been in a
state where running the standard formatter breaks the guard. The larger
multi-row fixtures (`HousekeepingFixture*` and the eval-note seeds) move to
`test_support_fixtures.rs` and are re-exported through `test_support`, so every
existing `crate::repositories::test_support::*` import path is unchanged.

`djinn-k8s`'s `launcher.rs` crossed the same ceiling once #2609's CAP_CHOWN work
and this branch's `warm_job_millicores` both landed. The CPU-quantity helpers
move to `launcher_cpu.rs` — a coherent unit rather than an arbitrary cut: one
place turns a rendered Kubernetes CPU quantity into the millicores that the
launcher's lease ceiling and the build-slot weight policy both reason about,
which is what keeps the quota a workload is charged and the CPU it is actually
given from drifting apart. Re-exported through `launcher`, so import paths are
unchanged there too.
